requirements.txt
aiohttp==3.5.4
beautifulsoup4==4.7.1
discord.py==1.2.3
